requirements.txt
python-telegram-bot
pydub
SpeechRecognition
openai==0.28 